+++ sakura_core/convert/CConvert_HankataToZenkata.cpp
@@ -18,7 +18,7 @@ bool CConvert_HankataToZenkata::DoConvert(CNativeW* pcData)
 {
 	//半角カナ→全角カナ
 	wchar_t* pBuf = new wchar_t[pcData->GetStringLength()+1]; //文字数が減ることはあっても増えることは無いので、これでＯＫ
-	int nDstLen = 0;
+	ssize_t nDstLen = 0;
 	Convert_HankataToZenkata(pcData->GetStringPtr(), pcData->GetStringLength(), pBuf, &nDstLen);
 	pcData->SetString(pBuf, nDstLen);
 	delete[] pBuf;
