+++ sakura_core/convert/CConvert_TabToSpace.h
@@ -13,7 +13,7 @@
 
 class CConvert_TabToSpace final : public CConvert{
 public:
-	CConvert_TabToSpace(int nTabWidth, int nStartColumn, bool bExtEol)
+	CConvert_TabToSpace(ssize_t nTabWidth, ssize_t nStartColumn, bool bExtEol)
 	: m_nTabWidth(nTabWidth), m_nStartColumn(nStartColumn), m_bExtEol(bExtEol)
 	{
 	}
@@ -21,8 +21,8 @@ public:
 	bool DoConvert(CNativeW* pcData) override;
 
 private:
-	int m_nTabWidth;
-	int m_nStartColumn;
+	ssize_t m_nTabWidth;
+	ssize_t m_nStartColumn;
 	bool m_bExtEol;
 };
 #endif /* SAKURA_CCONVERT_TABTOSPACE_DDBB0EE5_D2FB_4624_9A68_4DD4B81C0744_H_ */
